Environment mapping ^& Windows support

+++ src/utils/file.cpp
@@ -6,7 +6,7 @@
 std::vector<char> read_binary(std::string_view path) {
     std::ifstream file{path.data(), std::ios::ate | std::ios::binary};
     CHECK(file.is_open(), "");
-    long const file_size = file.tellg();
+    long long const file_size = file.tellg();
     std::vector<char> data((size_t) file_size);
     file.seekg(0);
     file.read(data.data(), file_size);
